+++ test/test-special-symbol.c
@@ -59,7 +59,7 @@ int is_bopomofo_collision_key( const char *key )
 		",",
 		".",
 	};
-	int i;
+	size_t i;
 
 	for ( i = 0; i < ARRAY_SIZE( COLLISION_KEY ); ++i ) {
 		if ( strcmp( key, COLLISION_KEY[i] ) == 0 ) {
@@ -72,7 +72,7 @@ int is_bopomofo_collision_key( const char *key )
 void test_in_chinese_mode()
 {
 	ChewingContext *ctx;
-	int i;
+	size_t i;
 
 	chewing_Init( NULL, NULL );
 
@@ -99,7 +99,7 @@ void test_in_chinese_mode()
 void test_in_easy_symbol_mode()
 {
 	ChewingContext *ctx;
-	int i;
+	size_t i;
 
 	chewing_Init( NULL, NULL );
 
@@ -135,7 +135,7 @@ int is_fullshape_collision_key( const char *key )
 		"+",
 		"-",
 	};
-	int i;
+	size_t i;
 
 	for ( i = 0; i < ARRAY_SIZE( COLLISION_KEY ); ++i ) {
 		if ( strcmp( key, COLLISION_KEY[i] ) == 0 ) {
@@ -148,7 +148,7 @@ int is_fullshape_collision_key( const char *key )
 void test_in_fullshape_mode()
 {
 	ChewingContext *ctx;
-	int i;
+	size_t i;
 
 	chewing_Init( NULL, NULL );
